gmap_hash_str_str_map code converage
